refactor: Introduce a function for filtering
Code to filter in the original code is duplicated for day and month
filter.  The process to take an input value until the user inputs one on
the list is common for both month and day filter.

Moved the functionality of filtering in to a specific function and then
call the function with month or day input lists.  This provide a cleaner
look on the filter code.  This function could also be re-used any where
else that a input value is entered until one is provided from the input
list

 - Function for filter added
 - Month filter called; duplicated code removed
 - Day filter called; duplicated code removed